# specs/openapi.yaml
openapi: 3.1.0
info:
  title: FlowProof API
  version: 0.6.0
paths:
...
  /api/v1/provider-connections/xero/status:
    get:
      tags:
      - providers
      summary: Xero Status
      operationId: xero_status_api_v1_provider_connections_xero_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Xero Status Api V1 Provider Connections Xero Status
                  Get
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/provider-connections/xero/pkce/start:
    post:
      tags:
      - providers
      summary: Xero Pkce Start
      operationId: xero_pkce_start_api_v1_provider_connections_xero_pkce_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/XeroPkceStartIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Xero Pkce Start Api V1 Provider Connections Xero Pkce
                  Start Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/provider-connections/xero/callback:
    get:
      tags:
      - providers
      summary: Xero Pkce Callback
      operationId: xero_pkce_callback_api_v1_provider_connections_xero_callback_get
      parameters:
      - name: state
        in: query
        required: true
        schema:
          type: string
          title: State
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: error
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/provider-connections/xero/disconnect:
    post:
      tags:
      - providers
      summary: Xero Disconnect
      operationId: xero_disconnect_api_v1_provider_connections_xero_disconnect_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Xero Disconnect Api V1 Provider Connections Xero Disconnect
                  Post
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/provider-connections/xero/qualify-invoice:
    post:
      tags:
      - providers
      summary: Xero Qualify Invoice
      operationId: xero_qualify_invoice_api_v1_provider_connections_xero_qualify_invoice_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/XeroQualificationIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Xero Qualify Invoice Api V1 Provider Connections Xero
                  Qualify Invoice Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    XeroPkceStartIn:
      properties:
        client_id:
          type: string
          maxLength: 128
          minLength: 8
          pattern: ^[A-Za-z0-9._-]+$
          title: Client Id
      type: object
      required:
      - client_id
      title: XeroPkceStartIn
    XeroQualificationIn:
      properties:
        invoice_id:
          type: string
          maxLength: 255
          minLength: 1
          title: Invoice Id
        expected_exists:
          type: boolean
          title: Expected Exists
          default: true
        expected_amount:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,16}|(?=[\d.]{1,19}0*$)\d{0,16}\.\d{0,2}0*$)
          title: Expected Amount
        currency:
          type: string
          pattern: ^[A-Z]{3}$
          title: Currency
        provider_identity_confirmed:
          type: boolean
          title: Provider Identity Confirmed
        entity_confirmed:
          type: boolean
          title: Entity Confirmed
        evidence_bounded:
          type: boolean
          title: Evidence Bounded
        no_mutation_confirmed:
          type: boolean
          title: No Mutation Confirmed
      type: object
      required:
      - invoice_id
      - expected_amount
      - currency
      - provider_identity_confirmed
      - entity_confirmed
      - evidence_bounded
      - no_mutation_confirmed
      title: XeroQualificationIn
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API credential issued to a human or service principal.
    SessionCookie:
      type: apiKey
      in: cookie
      name: flowproof_session
      description: Human browser session; state-changing requests also require X-CSRF-Token.